Refactor config parsing rules
The upshot is that the rules for parsing/extracting options from the
configuration file are now basically internal to the Config class.
The interface for adding rules to the config is much more argparse-like and
indeed simpler. The rules attribute has type Dict[str, Tuple[str, Closure]]
rather than Dict[str, SomeRule]. Is this better? Who knows?