Allow case-insensitive option keys
The default Config.parser allows for accessing config keys in a case-
insensitive manner. This commit extends that to Config's rule matching.
Consideration was given to make the matching for choices rules and actions
rules work the same way. After all, the --sort command-line option works this
way. However this was rejected until later.